# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Películas & Series Latino",
    "version": "1.0.0",
    "description": "API REST en vivo para consultar catálogos estilo Netflix, metadatos y reproductores embeds verificados de películas, series, animes y doramas en audio Latino."
  },
  "servers": [
    {
      "url": "https://api-pelis-series-latino.vercel.app",
      "description": "Servidor de Producción Vercel"
    }
  ],
  "paths": {
...
    "/api/v1/media/{id}": {
      "get": {
        "summary": "Obtener detalle por Slug o ID (metadata instantánea)",
        "description": "Devuelve la ficha completa (sinopsis, imágenes, reparto, temporadas) desde la base de datos o la caché, sin esperar a la resolución de enlaces. El bloque `streams` indica si los servidores ya están listos o dónde pedirlos (`GET /api/v1/media/{id}/streams`), lo que permite abrir una ficha emergente al instante y resolver la reproducción solo cuando el usuario pulsa Reproducir. Usa `streams=wait` para el comportamiento bloqueante clásico (metadata + servidores en una sola respuesta).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Slug o ID del contenido (ej. scary-movie-6, la-casa-del-dragon)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "streams",
            "in": "query",
            "required": false,
            "description": "wait | true | 1 → espera a que los servidores estén resueltos y los incluye en la respuesta (más lento). Por defecto la respuesta es instantánea y `streams.status` vale 'pending'.",
            "schema": {
              "type": "string",
              "enum": [
                "wait",
                "true",
                "1"
              ]
            }
          },
          {
            "name": "include",
            "in": "query",
            "required": false,
            "description": "season_1 | first_season → adjunta el primer episodio resuelto. streams → alias legacy de streams=wait.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle obtenido con éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media/{id}/streams": {
      "get": {
        "summary": "Resolver los enlaces reproducibles de un título",
        "description": "Camino lento aislado: resuelve los servidores del título (desde los enlaces persistidos si tienen menos de 24 h, si no scrapeando la fuente) y los guarda para las siguientes consultas. Pensado para llamarse al pulsar Reproducir, con la ficha ya pintada.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Slug o ID del contenido",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deep",
            "in": "query",
            "required": false,
            "description": "1 | true → fuerza la fusión multifuente completa (TioPlus + FuegoCine). Más lento; devuelve el máximo de servidores disponibles.",
            "schema": {
              "type": "string",
              "enum": [
                "1",
                "true"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Enlaces resueltos con éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamsResponse"
                }
              }
            }
          },
          "404": {
            "description": "El contenido solicitado no existe o no está disponible"
          }
        }
      }
    },
    "/api/v1/series/{id}": {
      "get": {
        "summary": "Obtener detalle de una serie (metadata instantánea)",
        "description": "Igual que /api/v1/media/{id} pero forzando el tipo serie: incluye temporadas y episodios. Los enlaces se piden en /api/v1/series/{id}/streams o por episodio.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Slug o ID de la serie",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "streams",
            "in": "query",
            "required": false,
            "description": "wait | true | 1 → espera a la resolución de servidores.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle obtenido con éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/series/{id}/streams": {
      "get": {
        "summary": "Resolver los enlaces reproducibles de una serie",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deep",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Enlaces resueltos con éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/series/{id}/season/{season}/episode/{episode}": {
      "get": {
        "summary": "Obtener servidores de streaming para un episodio específico de una serie",
        "description": "Devuelve los reproductores de video embed reales (VidHide, Netu, UPFAST, etc.) para la temporada y episodio indicados de la serie.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Slug de la serie (ej. la-casa-del-dragon)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "season",
            "in": "path",
            "required": true,
            "description": "Número de temporada (ej. 1)",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "episode",
            "in": "path",
            "required": true,
            "description": "Número de episodio (ej. 1)",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Servidores del episodio obtenidos con éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeDetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media/{id}/season/{season}/episode/{episode}": {
      "get": {
        "summary": "Obtener servidores de reproducción de un episodio de serie (Alias Media)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "season",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "episode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Servidores del episodio obtenidos con éxito"
          }
        }
      }
    },
...
    "/api/v1/media": {
      "get": {
        "summary": "Obtener catálogo general (películas + series)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "description": "Número de página real sobre el índice general, empezando en 1."
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            },
            "description": "Cantidad de ítems por página. Alias compatible: size. Máximo 100."
          },
          {
            "name": "size",
            "in": "query",
            "description": "Alias de limit para definir la cantidad de ítems por página. Máximo 100.",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "genre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Catálogo devuelto con éxito"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ServerOption": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "quality": {
            "type": "string",
            "example": "1080p"
          },
          "language": {
            "type": "string",
            "example": "latino"
          },
          "embed_url": {
            "type": "string",
            "example": "https://vidhideplus.com/v/f1kgebbi5bly"
          },
          "status": {
            "type": "string",
            "enum": [
              "online",
              "offline"
            ],
            "description": "Estado de salud del servidor verificado en la capa de aplicación (Soft Errors). Los reproductores o clientes DEBEN IGNORAR O FILTRAR los servidores con status: 'offline'."
          },
          "last_checked": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Episode": {
        "type": "object",
        "properties": {
          "episode_number": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Los herederos del Dragón"
          },
          "still_path": {
            "type": "string",
            "nullable": true
          },
          "servers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServerOption"
            }
          }
        }
      },
      "Season": {
        "type": "object",
        "properties": {
          "season_number": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Temporada 1"
          },
          "episodes_count": {
            "type": "integer",
            "example": 10
          },
          "episodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Episode"
            }
          }
        }
      },
      "MediaItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "la-casa-del-dragon"
          },
          "tmdb_id": {
            "type": "integer"
          },
          "type": {
            "type": "string",
            "enum": [
              "movie",
              "tvseries"
            ],
            "description": "Determina si el contenido es película o serie. Para 'movie', consultar `servers`. Para 'tvseries', consultar `seasons` o la ruta de episodios."
          },
          "title": {
            "type": "string"
          },
          "original_title": {
            "type": "string"
          },
          "overview": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          },
          "release_date": {
            "type": "string"
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "poster": {
            "type": "string",
            "nullable": true
          },
          "backdrop": {
            "type": "string",
            "nullable": true
          },
          "tagline": {
            "type": "string",
            "description": "Lema o eslogan oficial de TMDB"
          },
          "logo": {
            "type": "string",
            "nullable": true,
            "description": "Arte tipográfico del título (logo transparente de TMDB, es/en). Pensado para portadas y héroes."
          },
          "runtime": {
            "type": "integer",
            "nullable": true,
            "description": "Duración en minutos. En series, duración media del episodio."
          },
          "content_rating": {
            "type": "string",
            "nullable": true,
            "description": "Clasificación por edades oficial, priorizando mercados hispanohablantes (MX, AR, CL, ES, CO) y con US como respaldo."
          },
          "director": {
            "type": "string",
            "nullable": true,
            "description": "Director de la película."
          },
          "created_by": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            },
            "description": "Creadores de la serie."
          },
          "streams": {
            "$ref": "#/components/schemas/StreamsStatus"
          },
          "trailer": {
            "type": "string",
            "nullable": true,
            "description": "URL oficial de tráiler en YouTube (prioriza audio Latino)"
          },
          "cast": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Lista de nombres simples de actores del reparto (evita [object Object] en clientes)"
          },
          "cast_details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "character": {
                  "type": "string"
                },
                "photo": {
                  "type": "string",
                  "nullable": true
                }
              }
            },
            "description": "Lista detallada del reparto enriquecida desde TMDB con fotos y nombres de personaje"
          },
          "servers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServerOption"
            },
            "description": "Presente únicamente cuando type = movie"
          },
          "total_seasons": {
            "type": "integer",
            "description": "Presente únicamente cuando type = tvseries"
          },
          "total_episodes": {
            "type": "integer",
            "description": "Presente únicamente cuando type = tvseries"
          },
          "seasons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Season"
            },
            "description": "Lista de temporadas y episodios. Presente únicamente cuando type = tvseries"
          }
        }
      },
      "StreamsStatus": {
        "type": "object",
        "description": "Estado de los enlaces de un título. La metadata se devuelve siempre al instante; cuando status es 'pending' los servidores todavía no están resueltos y el cliente debe pedirlos a `url` (normalmente al pulsar Reproducir).",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "pending"
            ]
          },
          "url": {
            "type": "string",
            "example": "/api/v1/media/scary-movie-6/streams"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Momento de la última resolución de enlaces. Por debajo de 24 h se sirven directamente de la base de datos."
          }
        }
      },
      "StreamsResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "tmdb_id": {
                "type": "integer"
              },
              "type": {
                "type": "string",
                "enum": [
                  "movie",
                  "tvseries"
                ]
              },
              "title": {
                "type": "string"
              },
              "primary_stream": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ServerOption"
                  }
                ],
                "nullable": true,
                "description": "Servidor recomendado según prioridad de fuente, calidad e idioma."
              },
              "servers": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ServerOption"
                }
              },
              "seasons": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Season"
                },
                "description": "Presente únicamente en series."
              },
              "updated_at": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              }
            }
          }
        }
      },
...
      "MediaItemResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/MediaItem"
          }
        }
      },
      "EpisodeDetailResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "series_id": {
                "type": "string"
              },
              "season_number": {
                "type": "integer"
              },
              "episode_number": {
                "type": "integer"
              },
              "servers": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ServerOption"
                }
              }
            }
          }
        }
      }
    }
  }
}